requirements.txt
asgiref==3.5.2

Django==4.0.5
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
psycopg2>=2.8
backports.zoneinfo==0.2.1